Move helper functions to api_testing_common
